Move common http test utilities from testutils to common4j test fixtures
